+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/shard/commands/CommandParamInfoImpl.java
@@ -16,14 +16,14 @@ public class CommandParamInfoImpl implements CommandParamInfo {
 
     private List<String> tableNameList = Collections.emptyList(); // processed tables list
     private int commitBatchSize = DEFAULT_COMMIT_BATCH_SIZE;
-    private Long snapshotBlockHeight = -1L;
+    private Integer snapshotBlockHeight = -1;
     private byte[] shardHash;
     private Set<Long> dbIdExclusionSet;
 
     public CommandParamInfoImpl() {
     }
 
-    public CommandParamInfoImpl(List<String> tableNameList, int commitBatchSize, Long snapshotBlockHeight, Set<Long> dbIdExclusionSet) {
+    public CommandParamInfoImpl(List<String> tableNameList, int commitBatchSize, Integer snapshotBlockHeight, Set<Long> dbIdExclusionSet) {
         this.shardHash = new byte[0];
         this.tableNameList = tableNameList;
         this.commitBatchSize = commitBatchSize;
@@ -31,7 +31,7 @@ public class CommandParamInfoImpl implements CommandParamInfo {
         this.dbIdExclusionSet = dbIdExclusionSet == null ? Collections.emptySet() : dbIdExclusionSet;
     }
 
-    public CommandParamInfoImpl(List<String> tableNameList, int commitBatchSize, Long snapshotBlockHeight, byte[] shardHash, Set<Long> dbIdExclusionSet) {
+    public CommandParamInfoImpl(List<String> tableNameList, int commitBatchSize, Integer snapshotBlockHeight, byte[] shardHash, Set<Long> dbIdExclusionSet) {
         this(tableNameList, commitBatchSize, snapshotBlockHeight, dbIdExclusionSet);
         this.shardHash = shardHash;
     }
@@ -66,11 +66,11 @@ public class CommandParamInfoImpl implements CommandParamInfo {
         this.commitBatchSize = commitBatchSize;
     }
 
-    public Long getSnapshotBlockHeight() {
+    public Integer getSnapshotBlockHeight() {
         return snapshotBlockHeight;
     }
 
-    public void setSnapshotBlockHeight(Long snapshotBlockHeight) {
+    public void setSnapshotBlockHeight(Integer snapshotBlockHeight) {
         this.snapshotBlockHeight = snapshotBlockHeight;
     }
