+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/shard/commands/ReLinkDataCommand.java
@@ -26,11 +26,11 @@ public class ReLinkDataCommand implements DataMigrateOperation {
     private DataTransferManagementReceiver dataTransferManagement;
     private List<String> tableNameList;
     private int commitBatchSize;
-    private long snapshotBlockHeight;
+    private int snapshotBlockHeight;
     private Set<Long> dbIdsExclusionList;
 
     public ReLinkDataCommand(DataTransferManagementReceiver dataTransferManagement,
-                             int commitBatchSize, long snapshotBlockHeight, List<String> tableNameList, Set<Long> dbIdsExlusionList) {
+                             int commitBatchSize, int snapshotBlockHeight, List<String> tableNameList, Set<Long> dbIdsExlusionList) {
         this.dataTransferManagement = Objects.requireNonNull(dataTransferManagement, "dataTransferManagement is NULL");
         this.snapshotBlockHeight = snapshotBlockHeight;
         this.commitBatchSize = commitBatchSize <= 0 ? DEFAULT_COMMIT_BATCH_SIZE : commitBatchSize;
@@ -39,7 +39,7 @@ public class ReLinkDataCommand implements DataMigrateOperation {
     }
 
     public ReLinkDataCommand(DataTransferManagementReceiver dataTransferManagement,
-                             long snapshotBlockHeight, Set<Long> dbIdsExclusionList) {
+                             int snapshotBlockHeight, Set<Long> dbIdsExclusionList) {
         this(dataTransferManagement,  null, DEFAULT_COMMIT_BATCH_SIZE, snapshotBlockHeight, dbIdsExclusionList);
         //TODO move it to another class
         tableNameList.add(TRANSACTION_TABLE_NAME);
@@ -54,7 +54,7 @@ public class ReLinkDataCommand implements DataMigrateOperation {
             DataTransferManagementReceiver dataTransferManagement,
             List<String> tableNameList,
             int commitBatchSize,
-            Long snapshotBlockHeight, Set<Long> dbIdsExclusionList) {
+            int snapshotBlockHeight, Set<Long> dbIdsExclusionList) {
         this(dataTransferManagement, commitBatchSize, snapshotBlockHeight, tableNameList, dbIdsExclusionList);
     }
